pyproject.toml
[tool.poetry]
name = "slowlane"
version = "0.1.0"
description = "Python CLI tool for Apple service automation - fastlane-compatible authentication and App Store Connect/Developer Portal operations"
authors = ["Slowlane Contributors"]
license = "MIT"
readme = "README.md"
packages = [{include = "slowlane", from = "src"}]

[tool.poetry.scripts]
slowlane = "slowlane.cli.main:app"

[tool.poetry.dependencies]
python = "^3.11"
typer = "^0.21.1"
click = "^8.3.1"
httpx = "^0.28.1"
keyring = "^25.7.0"
cryptography = "^46.0.4"
pyjwt = "^2.11.0"
playwright = "^1.58.0"
rich = "^14.3.1"
tomli = "^2.4.0"
tomli-w = "^1.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
pytest-asyncio = "^1.3.0"
pytest-cov = "^7.0.0"
vcrpy = "^8.1.1"
mypy = "^1.19.1"
ruff = "^0.14.14"
types-toml = "^0.10.8.20240310"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM", "RUF"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = ["keyring.*", "playwright.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=src/slowlane --cov-report=term-missing --cov-report=html"

[tool.coverage.run]
source = ["src/slowlane"]
branch = true
omit = ["*/tests/*", "*/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
show_missing = true
